minor fixes
Former-commit-id: fba77c6bf0603c47b9332e7713254e9751c93fdc

+++ Code/Utilities/sscGraphicalPrimitives.cpp
@@ -38,7 +38,7 @@ GraphicalPoint3D::~GraphicalPoint3D()
 	}
 }
 
-void GraphicalPoint3D::setRadius(int radius)
+void GraphicalPoint3D::setRadius(double radius)
 {
 	source->SetRadius(radius);
 }
